Move panel layout chrome out of main.ts
The measure-bar and dock clearance observers, the rail collapse toggle and
its chevrons, and the panel wheel containment are panel geometry, not app
wiring: they read element boxes and write CSS custom properties, and touch
no viewer or scan state. They now live in src/ui/panelChrome.ts.

Pure move. main.ts drops 7515 -> 7360 lines, so the shrink-only ratchet is
green again and the baseline is banked at the lower count. Architecture map
and the v0.6.1 limitations note follow the new count; the unsafeHtml
allowlist entry for the chevron follows the code to its new file.
