+++ tests/unsafeHtmlGuard.test.ts
@@ -123,9 +123,9 @@ const ALLOWLIST: readonly string[] = [
   'src/ui/ExportPanel.ts::Export / Convert',
   'src/ui/FullscreenToggle.ts::ICON_ENTER',
   // P11 left-rail toggle: `chevron` is a hardcoded, literal static SVG string
-  // (a chevron path) defined inline in wireLeftRailToggle — no user data, same
+  // (a chevron path) defined inline in panelChrome.ts — no user data, same
   // sanctioned pattern as the other icon SVGs above.
-  'src/main.ts::cfg.chevron',
+  'src/ui/panelChrome.ts::cfg.chevron',
   // Colorbar legend overlay (hand-verified 2026-07-14). `buildColorbarSvg` is
   // the pure figure-legend generator: it XML-escapes EVERY text value it
   // interpolates (see `esc()` in src/render/colorbar.ts, pinned by
